+++ src/thor/triplegbuilder.cc
@@ -1344,10 +1344,10 @@ TripLeg_Edge* AddTripEdge(const AttributesController& controller,
   }
 
   if (directededge->access_restriction() && edge_itr->restriction_index != kInvalidRestriction) {
-    const std::vector<baldr::AccessRestriction>& restrictions =
-        graphtile->GetAccessRestrictions(edge.id(), costing->access_mode());
-    trip_edge->mutable_restriction()->set_type(
-        static_cast<uint32_t>(restrictions[edge_itr->restriction_index].type()));
+    auto restriction = graphtile->GetAccessRestrictionAtIndex(edge.id(), costing->access_mode(),
+                                                              edge_itr->restriction_index);
+    assert(restriction != nullptr);
+    trip_edge->mutable_restriction()->set_type(static_cast<uint32_t>(restriction->type()));
   }
 
   trip_edge->set_has_time_restrictions(edge_itr->restriction_index != kInvalidRestriction);